# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation for API e-mosque endpoints",
        "version": "1.0.0"
    },
    "paths": {
        "/api/saldo-kas": {
            "get": {
                "summary": "Read all kas",
                "description": "Read all kas data.",
                "operationId": "readAllKas",
                "responses": {
                    "200": {
                        "description": "Successful retrieval of kas data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "description": "Message indicating the status of the kas retrieval",
                                            "type": "string",
                                            "example": "Data saldo kas berhasil didapat"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No kas data available",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "description": "Message indicating the status of the kas retrieval failure",
                                            "type": "string",
                                            "example": "Data kas tidak tersedia"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create new kas",
                "description": "Create a new kas with total saldo, kas jenis ID, and current date.",
                "operationId": "createKas",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "total_saldo",
                                    "kas_jenis_id"
                                ],
                                "properties": {
                                    "total_saldo": {
                                        "description": "Total saldo of the kas",
                                        "type": "number",
                                        "example": 1000000
                                    },
                                    "kas_jenis_id": {
                                        "description": "ID of the kas jenis",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful creation of kas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "description": "Message indicating the status of the kas creation",
                                            "type": "string",
                                            "example": "Inputan data saldo kas berhasil dibuat"
                                        },
                                        "data": {
                                            "description": "Created kas data",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "description": "Message indicating the status of the kas creation failure",
                                            "type": "string",
                                            "example": "Inputan data saldo kas gagal"
                                        },
                                        "data": {
                                            "description": "Additional error data",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/saldo-kas/{id_kas}": {
            "put": {
                "summary": "Update kas",
                "description": "Update kas data by ID.",
                "operationId": "updateKas",
                "parameters": [
                    {
                        "name": "id_kas",
                        "in": "path",
                        "description": "ID of the kas to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "total_saldo": {
                                        "description": "Updated total saldo of the kas",
                                        "type": "number",
                                        "example": 2000000
                                    },
                                    "kas_jenis_id": {
                                        "description": "Updated ID of the kas jenis",
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful update of kas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "description": "Message indicating the status of the kas update",
                                            "type": "string",
                                            "example": "Saldo kas berhasil diperbarui"
                                        },
                                        "data": {
                                            "description": "Updated kas data",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Kas data not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "description": "Message indicating the status of the kas update failure",
                                            "type": "string",
                                            "example": "Data saldo kas tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "summary": "Delete kas",
                "description": "Delete kas data by ID.",
                "operationId": "deleteKas",
                "parameters": [
                    {
                        "name": "id_kas",
                        "in": "path",
                        "description": "ID of the kas to be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful deletion of kas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "description": "Message indicating the status of the kas deletion",
                                            "type": "string",
                                            "example": "Saldo kas berhasil dihapus"
                                        },
                                        "data": {
                                            "description": "No data returned",
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Kas data not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "description": "Message indicating the status of the kas deletion failure",
                                            "type": "string",
                                            "example": "Data saldo kas tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/saldo-kas/{id_kas}/approve": {
            "get": {
                "summary": "Approve kas",
                "description": "Approve kas data by ID.",
                "operationId": "approveKas",
                "parameters": [
                    {
                        "name": "id_kas",
                        "in": "path",
                        "description": "ID of the kas to be approved",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "approval",
                        "in": "query",
                        "description": "Approval status ('true' for approve, 'false' for disapprove)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "true",
                                "false"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful approval of kas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "description": "Message indicating the status of the kas approval",
                                            "type": "string",
                                            "example": "Saldo kas berhasil disetujui"
                                        },
                                        "data": {
                                            "description": "Approved kas data",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Kas data not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "description": "Message indicating the status of the kas approval failure",
                                            "type": "string",
                                            "example": "Data saldo kas tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}